BugFix:  Unable to move pin when using Annotation? or MKAnnotation? as the class-wide variable to store positional values prior to move. Solution: Used CLLocationCoordinate2D?
When assigning a value to class-wide Annotation/MKKAnnotation variable, it changes as the original source variable change.  So after move, the value is new positional value.  Old/original latitude/longitude prior to start/dragging is gone.
Looks like they point to same memory.  So it's not a standard x = y assignment.
	Ex:
		Changing x afterwards, also changes y.