CoreData changes made.
Pin Entity --> now has photoCount entry.  This was created to FlickrController will know the expected number of photos incoming (in-case it's still coming down asynchronously over network).  And adjust AI accordingly.
Photo Entity --> URLString was added earlier but not utilized.  Now when photo is saved as Data, its corresponding URL is also saved.  For now the thinking is that this URL can be manipulated via string functions to retrieve same image with different Quality.  Or perhaps reload the picture later by any VC that has the photo entity and something is wrong with the imageData it pulled up from coreData