Background context used at each stage when saving to CoreData
func downloadNearbyPhotosToPin()
	* Uses the backgroundContext defined in DataContoller.init()
		* Requires an explicit unwrap that should never fail because its value/creation is hard-coded inside init file

func connectPhotoAndPin()
	* Uses dataController.persistentContainer.performBackgroundTask to enable parallel processing if iOS deems performance advantage
		* temporary background context created for each iteration

Added NotificationObservers to help Main and Background on DataController communicate with each other.  For now, there's nothing to add feature wise because the NSFetchResults Delegate is updating UI.  This is more of a placeholder for future implementation and might need to be deleted because it provides no additional functionality later.